Only echo unset for ENV vars
Prior to this commit, the Make clean function would run 'unset' for the
environment variables used by this bot, but they wouldn't actually be
unset since the Makefile spawned a new shell env. Now it just echos the
commands to stdout and you can copypasta them. It might not be any
better.